+++ Sources/StreamChat/APIClient/ChatRemoteNotificationHandler.swift
@@ -16,16 +16,6 @@ public class MessageNotificationContent {
     }
 }
 
-public class ReactionNotificationContent {
-    public let message: ChatMessage
-    public let channel: ChatChannel?
-
-    public init(message: ChatMessage, channel: ChatChannel?) {
-        self.message = message
-        self.channel = channel
-    }
-}
-
 public class UnknownNotificationContent {
     public let content: UNNotificationContent
 
@@ -36,7 +26,6 @@ public class UnknownNotificationContent {
 
 public enum ChatPushNotificationContent {
     case message(MessageNotificationContent)
-    case reaction(ReactionNotificationContent)
     case unknown(UnknownNotificationContent)
 }
 
@@ -84,6 +73,7 @@ public class ChatRemoteNotificationHandler {
     let database: DatabaseContainer
     let syncRepository: SyncRepository
     let messageRepository: MessageRepository
+    let extensionLifecycle: NotificationExtensionLifecycle
 
     public init(client: ChatClient, content: UNNotificationContent) {
         self.client = client
@@ -91,6 +81,7 @@ public class ChatRemoteNotificationHandler {
         database = client.databaseContainer
         syncRepository = client.syncRepository
         messageRepository = client.messageRepository
+        extensionLifecycle = client.extensionLifecycle
     }
 
     public func handleNotification(completion: @escaping (ChatPushNotificationContent) -> Void) -> Bool {
@@ -130,7 +121,11 @@ public class ChatRemoteNotificationHandler {
 
     private func getMessageAndSync(cid: ChannelId, messageId: String, completion: @escaping (ChatMessage?, ChatChannel?) -> Void) {
         let database = self.database
-        messageRepository.getMessage(cid: cid, messageId: messageId) { [weak self] result in
+        messageRepository.getMessage(
+            cid: cid,
+            messageId: messageId,
+            store: !extensionLifecycle.isAppReceivingWebSocketEvents
+        ) { [weak self] result in
             guard case let .success(message) = result else {
                 completion(nil, nil)
                 return