+++ Sources/StreamChat/Database/DTOs/ChannelDTO.swift
@@ -333,7 +333,8 @@ extension NSManagedObjectContext {
 
 extension ChannelDTO {
     static func channelListFetchRequest(
-        query: ChannelListQuery
+        query: ChannelListQuery,
+        chatClientConfig: ChatClientConfig
     ) -> NSFetchRequest<ChannelDTO> {
         let request = NSFetchRequest<ChannelDTO>(entityName: ChannelDTO.entityName)
 
@@ -351,9 +352,13 @@ extension ChannelDTO {
         // We can't pass bools directly to NSPredicate so we have to use integers
         let isHidden = NSPredicate(format: "isHidden == %i", query.filter.hiddenFilterValue == true ? 1 : 0)
 
-        let subpredicates = [
+        var subpredicates: [NSPredicate] = [
             matchingQuery, notDeleted, isHidden
         ]
+
+        if chatClientConfig.isChannelAutomaticFilteringEnabled, let filterPredicate = query.filter.predicate {
+            subpredicates.append(filterPredicate)
+        }
 
         request.predicate = NSCompoundPredicate(type: .and, subpredicates: subpredicates)
         request.fetchLimit = query.pagination.pageSize