+++ TestTools/StreamChatTestTools/Mocks/StreamChat/Repositories/MessageRepository_Mock.swift
@@ -12,10 +12,10 @@ final class MessageRepository_Mock: MessageRepository, Spy {
     }
 
     var sendMessageResult: Result<ChatMessage, MessageRepositoryError>?
-    var sendMessageCalls: [MessageId: (Result<ChatMessage, MessageRepositoryError>) -> Void] = [:]
+    @Atomic var sendMessageCalls: [MessageId: (Result<ChatMessage, MessageRepositoryError>) -> Void] = [:]
     var getMessageResult: Result<ChatMessage, Error>?
+    var receivedGetMessageStore: Bool?
     var saveSuccessfullyDeletedMessageError: Error?
-    let lock = NSLock()
     var updatedMessageLocalState: LocalMessageState?
 
     override func sendMessage(
@@ -23,11 +23,11 @@ final class MessageRepository_Mock: MessageRepository, Spy {
         completion: @escaping (Result<ChatMessage, MessageRepositoryError>) -> Void
     ) {
         record()
-        lock.lock()
-        sendMessageCalls[messageId] = { result in
-            completion(result)
+        _sendMessageCalls.mutate { dictionary in
+            dictionary[messageId] = { result in
+                completion(result)
+            }
         }
-        lock.unlock()
 
         if let sendMessageResult = sendMessageResult {
             completion(sendMessageResult)
@@ -48,8 +48,9 @@ final class MessageRepository_Mock: MessageRepository, Spy {
         completion()
     }
 
-    override func getMessage(cid: ChannelId, messageId: MessageId, completion: ((Result<ChatMessage, Error>) -> Void)? = nil) {
+    override func getMessage(cid: ChannelId, messageId: MessageId, store: Bool = true, completion: ((Result<ChatMessage, Error>) -> Void)? = nil) {
         record()
+        receivedGetMessageStore = store
         getMessageResult.map { completion?($0) }
     }